+++ nym-vpn-app/src/i18n/config.ts
@@ -35,6 +35,8 @@ export const languages = [
   { code: 'es', name: 'Español' },
   { code: 'fr', name: 'Français' },
   { code: 'hi', name: 'हिन्दी' },
+  { code: 'id', name: 'Bahasa Indonesia' },
+  { code: 'it', name: 'Italiano' },
   { code: 'pt', name: 'Português Brasileiro' },
   { code: 'ru', name: 'Русский язык' },
   { code: 'tr', name: 'Türkçe' },
@@ -46,7 +48,6 @@ export const languages = [
   // { code: 'cs', name: 'Čeština (Czech)' },
   // { code: 'hu', name: 'Magyar (Hungarian)' },
   // { code: 'el', name: 'ελληνικά' },
-  // { code: 'it', name: 'Italiano' },
   // { code: 'ja', name: '日本語' },
 ] as const;